+++ src/main/java/com/modulo/functions/CubeFunction.java
@@ -3,7 +3,18 @@ package com.modulo.functions;
 import com.modulo.internal.CalcFunction;
 
 public class CubeFunction implements CalcFunction {
-    @Override public String getName() { return "cube"; }
-    @Override public String getInsertText() { return "cube("; }
-    @Override public double evaluate(double x) { return x * x * x; }
+    @Override
+    public String getName() {
+        return "cube";
+    }
+
+    @Override
+    public String getInsertText() {
+        return "cube(";
+    }
+
+    @Override
+    public double execute(double... args) {
+        return Math.pow(args[0], 3);
+    }
 }